+++ app/src/pages/Sales/Sales.js
@@ -5,39 +5,38 @@ import Spinner from '../../components/Spinner/Spinner';
 import SalesData from './SalesData';
 import './Sales.scss';
 import { connect } from 'react-redux';
-import Api from '../../redux/Api';
-import { waitSalesPage, waitSalesPageEnd } from '../../redux/actions';
+import { waitSalesPage, waitSalesPageEnd } from '../../redux/actions/sales';
 
 class Sales extends Component{
 
     static propTypes = {
         spinner: PropTypes.bool,
-        data: PropTypes.object
-    }
+        data: PropTypes.array
+    };
 
 
     componentDidMount(){
         this.props.waitSalesPage();
     }
-
     render(){
         return ( this.props.data ?
-          <div className='sales-wrap'>
-            { SalesData.map( item =>
-              <SalesItem key={item.id}
-                 salesHeadline={item.salesHeadline}
-                 content={item.content}
-                 expireDate={item.expireDate}
-                 salesImgCaption={item.salesImgCaption}
-              />) }
-          </div> : <Spinner />
+                <div className='sales-wrap'>
+                    { SalesData.map( item =>
+                        <SalesItem key={item.id}
+                                   salesHeadline={item.salesHeadline}
+                                   content={item.content}
+                                   expireDate={item.expireDate}
+                                   salesImgCaption={item.salesImgCaption}
+                        />) }
+                </div> : <Spinner />
         );
+
     }
 }
 
 const mapStateToProps = state => ({
-        spinner: state.spinner.spinner,
-        data: state.spinner.data
+        sales: state.sales.sales,
+        data: state.sales.data
      /*   showMessage: state.sendDataSpinner.showMessage*/
     });
