--- app/src/redux/constants/index.js
@@ -1,7 +1,0 @@
-const ActionTypes = {
-    PAGE_SALES_WAIT: 'PAGE_SALES_WAIT',
-    PAGE_SALES_SUCCESS: 'PAGE_SALES_SUCCESS',
-    ERROR: 'ERROR'
-};
-
-export default ActionTypes;
