+++ app/src/redux/middleware/index.js
@@ -1,5 +1,5 @@
 import { all } from 'redux-saga/effects';
-import sendData from './sendData';
+import sendData from './Sales/sales';
 
 export default function* rootSaga() {
     yield all([
