+++ app/src/redux/reducers/Sales/sales.js
@@ -1,21 +1,17 @@
-import ActionTypes from '../constants';
+import { PAGE_SALES_WAIT, PAGE_SALES_SUCCESS } from '../../constants/sales';
 
 const initialState = {
     spinner: false,
-    data: null,
-    error: null
+    data: null
 };
 
 export default (state = initialState, action) => {
     switch (action.type) {
-        case ActionTypes.PAGE_SALES_WAIT:
+        case PAGE_SALES_WAIT:
             return { ...state, spinner: true };
 
-        case ActionTypes.PAGE_SALES_SUCCESS:
+        case PAGE_SALES_SUCCESS:
             return { ...state, spinner: false, data: action.data };
-
-        case ActionTypes.ERROR:
-            return { ...state, spinner: false, error:false};
 
         default:
             return state;
